Refactor SAS URL generation into a separate helper class and streamline image URL handling in GetAlpakaFunction and GetAlpakasFunction.
