+++ Api/GetAlpakasFunction.cs
@@ -1,7 +1,6 @@
 using Azure.Data.Tables;
 using Azure.Storage;
 using Azure.Storage.Blobs;
-using Azure.Storage.Sas;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 using Microsoft.Extensions.Logging;
@@ -26,38 +25,17 @@ public class GetAlpakasFunction(
         TableClient tableClient = _tableServiceClient.GetTableClient("alpakas");
         BlobContainerClient container = _blobServiceClient.GetBlobContainerClient("alpakas");
 
+        var storageAccountName = Environment.GetEnvironmentVariable(EnvironmentVariables.StorageAccountName)
+            ?? throw new InvalidOperationException("Environment variable 'AZURE_STORAGE_ACCOUNT_NAME' is not set.");
+        var storageAccountKey = Environment.GetEnvironmentVariable(EnvironmentVariables.StorageAccountKey)
+            ?? throw new InvalidOperationException("Environment variable 'AZURE_STORAGE_ACCOUNT_KEY' is not set.");
+
         var alpakas = tableClient
             .Query<AlpakaEntity>()
             .OrderBy(a => a.Name)
             .Select(a =>
             {
-                string? sasUrl = null;
-                if (!string.IsNullOrWhiteSpace(a.ImageUrl))
-                {
-                    string blobName = Path.GetFileName(new Uri(a.ImageUrl).AbsolutePath);
-                    BlobClient blob = container.GetBlobClient(blobName);
-
-                    var expiresOn = DateTimeOffset.UtcNow.AddMinutes(30);
-                    var sasBuilder = new BlobSasBuilder
-                    {
-                        BlobContainerName = container.Name,
-                        BlobName = blobName,
-                        Resource = "b",
-                        ExpiresOn = expiresOn
-                    };
-                    sasBuilder.SetPermissions(BlobSasPermissions.Read);
-
-                    var storageAccountName = Environment.GetEnvironmentVariable(EnvironmentVariables.StorageAccountName)
-                        ?? throw new InvalidOperationException("Environment variable 'AZURE_STORAGE_ACCOUNT_NAME' is not set.");
-                    var storageAccountKey = Environment.GetEnvironmentVariable(EnvironmentVariables.StorageAccountKey) 
-                        ?? throw new InvalidOperationException("Environment variable 'AZURE_STORAGE_ACCOUNT_KEY' is not set.");
-                    StorageSharedKeyCredential credential = new(
-                        storageAccountName,
-                        storageAccountKey
-                    );
-                    var sasToken = sasBuilder.ToSasQueryParameters(credential).ToString();
-                    sasUrl = $"{blob.Uri}?{sasToken}";
-                }
+                string? sasUrl = SasTokenHelper.GenerateSasUrl(container, a.ImageUrl, storageAccountName, storageAccountKey);
 
                 return new
                 {
